# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bizny API - Productivity and venture coordination platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      operationId: listFeed
      tags: [feed]
      summary: Get unified feed posts
      parameters:
        - name: postType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [feed]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /feed/{id}/react:
    post:
      operationId: reactToPost
      tags: [feed]
      summary: React to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /feed/{id}/comments:
    get:
      operationId: listPostComments
      tags: [feed]
      summary: List comments on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
    post:
      operationId: createPostComment
      tags: [feed]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
  /feed/{id}/repost:
    post:
      operationId: repostPost
      tags: [feed]
      summary: Repost a post (direct or with quote)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepostInput"
      responses:
        "201":
          description: Created repost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /feed/{id}/save:
    post:
      operationId: savePost
      tags: [feed]
      summary: Save a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveStatus"
    delete:
      operationId: unsavePost
      tags: [feed]
      summary: Unsave a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveStatus"
...
  /users/{id}/request-verification:
    post:
      operationId: requestVerification
      tags: [fieldAgent]
      summary: Request field agent verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRequest"
...
  /verification-requests/{id}/status:
    patch:
      operationId: updateVerificationStatus
      tags: [fieldAgent]
      summary: Update verification request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRequest"
...
components:
  schemas:
...
    Post:
      type: object
      required: [id, content, author, postType, loves, unsures, comments, shares, saves, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserMini"
        postType:
          type: string
          enum: [share, question, request, opportunity, template, venture_progress, milestone, marketplace_listing, industry_insight, verification_update, innovation_idea, innovation_challenge]
        ventureId:
          type: ["integer", "null"]
        templateId:
          type: ["integer", "null"]
        mediaUrl:
          type: ["string", "null"]
        loves:
          type: integer
        unsures:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        saves:
          type: integer
        progressPercent:
          type: ["number", "null"]
        milestoneTag:
          type: ["string", "null"]
        requestCategory:
          type: ["string", "null"]
        locationName:
          type: ["string", "null"]
        locationCoords:
          type: ["string", "null"]
        stateCity:
          type: ["string", "null"]
        localArea:
          type: ["string", "null"]
        mainIndustry:
          type: ["string", "null"]
        subIndustry:
          type: ["string", "null"]
        activityTag:
          type: ["string", "null"]
        valueChainStage:
          type: ["string", "null"]
        mediaUrls:
          type: array
          items:
            type: string
          nullable: true
        linkUrl:
          type: ["string", "null"]
        userReaction:
          type: ["string", "null"]
        repostOf:
          type: ["integer", "null"]
        repostComment:
          type: ["string", "null"]
        reposts:
          type: integer
        isSaved:
          type: boolean
        originalPost:
          $ref: "#/components/schemas/Post"
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [content, postType]
      properties:
        content:
          type: string
        postType:
          type: string
          enum: [share, question, request, opportunity, template, venture_progress, milestone, marketplace_listing, industry_insight, verification_update, innovation_idea, innovation_challenge]
        ventureId:
          type: integer
        templateId:
          type: integer
        mediaUrl:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        linkUrl:
          type: string
        milestoneTag:
          type: string
        requestCategory:
          type: string
        progressPercent:
          type: number
        locationName:
          type: string
        locationCoords:
          type: string
        stateCity:
          type: string
        localArea:
          type: string
        mainIndustry:
          type: string
        subIndustry:
          type: string
        activityTag:
          type: string
        valueChainStage:
          type: string
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [love, unsure]
    PostComment:
      type: object
      required: [id, postId, content, author, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserMini"
        createdAt:
          type: string
          format: date-time
    PostCommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...
    RepostInput:
      type: object
      properties:
        comment:
          type: string
    SaveStatus:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
    UserMini:
      type: object
      required: [id, name, role, verificationStatus]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        verificationStatus:
          type: string
        avatarUrl:
          type: ["string", "null"]
...
    VerificationRequest:
      type: object
      required: [id, requesterId, status, targetType, createdAt]
      properties:
        id:
          type: integer
        requesterId:
          type: integer
        agentId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, in_review, verified, rejected]
        notes:
          type: ["string", "null"]
        evidenceUrls:
          type: array
          items:
            type: string
        targetType:
          type: string
        targetId:
          type: ["integer", "null"]
        rejectionReason:
          type: ["string", "null"]
        requester:
          $ref: "#/components/schemas/UserMini"
        createdAt:
          type: string
          format: date-time
    VerificationRequestInput:
      type: object
      required: [targetType]
      properties:
        notes:
          type: string
        targetType:
          type: string
        targetId:
          type: integer
    VerificationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, in_review, verified, rejected]
        notes:
          type: string
        evidenceUrls:
          type: array
          items:
            type: string
        rejectionReason:
          type: string
...